expense-tracker-env/requirements.txt
Django>=3.x
djangorestframework
django-cors-headers